Work4/submission/requirements.txt
ace_tools==0.0
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
importlib_resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.43.0
matplotlib==3.9.2
numba==0.60.0
numpy==2.0.2
packaging==24.1
pandas==2.2.3
patsy==0.5.6
pillow==11.0.0
pynndescent==0.5.13
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scikit-posthocs==0.10.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sklearn-relief==1.0.0b2
statsmodels==0.14.4
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2024.2
umap-learn==0.5.7
zipp==3.20.2
